Exract and fix public metadata tests
